+++ VersionHistory.md
@@ -6,7 +6,7 @@ Add changes here when they're committed to the `master` branch. Move them to "Re
 is updated in preparation for publishing an updated NuGet package.
 
 Prefix the description of the change with `[major]`, `[minor]` or `[patch]` in accordance with [SemVer](http://semver.org).
-
+* [minor] Support responses with an explicit `StatusCode` property.
 * [minor] Add `WebServiceRequestSettings.StartTrace` callback.
 
 ## Released